[move] unwrapOptionTResourceをResourceExtraに移動する
